First Steps
The initial importing is done. So far the progress that has been made is that the script can launch google chrome and open a google search page. There is a warning being thrown but it does not seem to impact the functionality. The next step is to have the script conduct searches using the sample information about the target person.