Updates to instagramSearch() and built dictionaryParser
Worked more on the instagram search function, it now can search for a particular name, select the first profile, and search to see if its bio contains any of the targets attributes. The next steps for this function is to have search through the first 10 profiles (if possible). Additionally the dictionaryParser was created to parse through the inputDictionary from the target for the search functions. Overall next steps are to implement a class for the target, implement the scoring system and implement the linkedin search function.