Refinement of instragramSearch()
The instagramSearch() method can now successfully search through as many search results as specified by the user, with the caveat being that searching through more results, the longer the program will take to run. Next step is to look at allowing the program to check for abbreviations of things one may find in the bio of the target (i.e. abbreviation of a university name)