Refined Target and Result Classes
Instances of the Target class can now be created using simple strings and a list as the constructor was changed. Additionally it now can store the results from the instagram search function, sort them by descending score and display them. The next steps could be to add printing the results at the end of the instragram search function and begin working on the linkedin search function.